refactor: create a logical structure for app components
This will probably change again in the future but we have to start with what we have :)

Closes: https://github.com/zakodium-oss/analysis-ui-components/issues/224
