--- src/app/components/index.ts
@@ -1,7 +1,0 @@
-export * from './IRPeaksPanel';
-export * from './MeasurementPlot';
-export * from './MeasurementMassPlot';
-export * from './MeasurementExplorer';
-export * from './MeasurementsPanel';
-export * from './MeasurementInfoPanel';
-export * from './SignalProcessingPanel';
