+++ src/app/index.ts
@@ -1,1 +1,4 @@
-export * from './components/index';
+export * from './explorer/index';
+export * from './kinds/ir/index';
+export * from './kinds/mass/index';
+export * from './panels/index';
