+++ src/app/kinds/ir/index.ts
@@ -1,0 +1,1 @@
+export * from './IRPeaksPanel';
