+++ src/app/kinds/mass/index.ts
@@ -1,0 +1,1 @@
+export * from './MeasurementMassPlot';
