+++ src/app/panels/MeasurementInfoPanel.tsx
