+++ src/app/panels/MeasurementsPanel.tsx
