+++ src/app/panels/SignalProcessingPanel.tsx
