+++ src/app/panels/index.ts
@@ -1,0 +1,3 @@
+export * from './MeasurementInfoPanel';
+export * from './MeasurementsPanel';
+export * from './SignalProcessingPanel';
